apps/ml-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
scikit-learn==1.6.0
prophet==1.1.6
numpy==2.2.0
scipy==1.15.0
psycopg2-binary==2.9.10
redis==5.2.0
pydantic==2.10.0
python-dotenv==1.0.1
httpx==0.28.0
